requirements.txt
astroid==2.4.1
backpack==0.1
blinker==1.4
cleo==0.6.8
click==7.1.2
colorama==0.4.3
Faker==0.8.18
Flask==1.1.2
inflection==0.3.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
orator==0.9.9
pastel==0.1.1
pendulum==1.5.1
psycopg2==2.8.5
pyaml==16.12.2
Pygments==2.6.1
pylev==1.3.0
pylint==2.5.2
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
pytzdata==2019.3
PyYAML==5.3.1
simplejson==3.17.0
six==1.15.0
text-unidecode==1.2
toml==0.10.1
typed-ast==1.4.1
tzlocal==1.5.1
Werkzeug==1.0.1
wrapt==1.12.1
